reorganize decode funtion into baes_head
